Merge backup

+++ backend/controllers/shiftsController.js
@@ -1,19 +1,26 @@
 // import model for use in controller functions
 const { Shift } = require('../models/Shift')
-// TODO: Add authorize middleware and checkManager/ checkEmployee middleware to all these routes
+const { Business } = require('../models/Business')
+const { Employee } = require('../models/Employee')
 
 // Logic to create shift
-const createShift = (req, res) => {
+const createShift = async (req, res) => {
   // I'm put everything in a try-catch block because I'm paranoid
   try {
     const {date, location, startTime, endTime,
       standardMinutes, overtimeMinutes, doubleTimeMinutes, totalPay} = req.body
 
-    // 1. Get the user Id from the jwt payload
-    const userId = '5b53377c46556409ebbad3c5' // TODO: Change this to userId = req.user._id after the authorize middleware has been added
-
-    // 2. Get the business Id from the jwt payload
-    const businessId = '5b53377c46556409ebbad3bc' // TODO: Change this to businessId = req.user.businessId after the authorize middleware has been added
+    let userId = ''
+    let businessId = ''
+    if (process.env.NODE_ENV === 'development') {
+      let employeesList = await Employee.find()
+      userId = employeesList[0]._id
+      let businesses = await Business.find()
+      businessId = businesses[0]._id // This is only to allow for development 
+    } else {
+      userId = req.user._id
+      businessId = req.user.businessId
+    }
 
     // 4. Create new shift object
     const shiftJson = {
@@ -30,17 +37,24 @@ const createShift = (req, res) => {
       business: businessId
     }
 
-    // 4. Save new shift
     const newShift = new Shift(shiftJson)
-    newShift.save()
-      .then(newShift => {
-        return res.status(200).json(newShift)
-      })
-      .catch(err => {
-        return res.status(400).json({
-          err: err.message
+
+    let isDemo
+    if (req.user) { isDemo = req.user.isDemo }
+    if (isDemo) { // If the user is a demo, return a success response without updating the database
+      return res.status(200).json(newShift)
+    } else {
+      // 4. Save new shift
+      newShift.save()
+        .then(newShift => {
+          return res.status(200).json(newShift)
         })
-      })
+        .catch(err => {
+          return res.status(400).json({
+            err: err.message
+          })
+        })
+    }
   } catch (error) {
     res.status(500).send('Something went wrong')
   }
@@ -50,8 +64,14 @@ const createShift = (req, res) => {
 const getEmployeeShifts = async (req, res) => {
   // I'm put everything in a try-catch block because I'm paranoid
   try {
+    let userId = ''
     // 1. Get the user Id from the jwt payload
-    const userId = '5b53377c46556409ebbad3c5' // TODO: Change this to userId = req.user._id after the authorize middleware has been added
+    if (process.env.NODE_ENV === 'development') {
+      let employeesList = await Employee.find()
+      userId = employeesList[0]._id // This is only to allow for development
+    } else {
+      userId = req.user._id
+    }
 
     // 2. Fetch all the shifts where the 'employee' property matches the Id
     const shifts = await Shift.find({ employee: userId })
@@ -85,17 +105,23 @@ const archiveShift = async (req, res) => {
       return res.status(404).send('Shift not found')
     }
 
-    // 4. If found, update the shift's 'status' propserty to 'archived'
-    shift.set({
-      status: 'archived'
-    })
+    let isDemo
+    if (req.user) { isDemo = req.user.isDemo }
+    if (isDemo) { // If the user is a demo, return a success response without updating the database
+      shift.status = 'archived'
+      return res.status(200).send(shift)
+    } else {
+      // 4. If found, update the shift's 'status' propserty to 'archived'
+      shift.set({
+        status: 'archived'
+      })
 
-    // 5. Save the update to the database
-    const result = await shift.save()
+      // 5. Save the update to the database
+      const result = await shift.save()
 
-    // 6. Send back the saved shift
-    res.send(result) // Console log the updated movie
-
+      // 6. Send back the saved shift
+      return res.status(200).send(result)
+    }
   } catch (error) {
     // An error will be thrown if the shiftId is not a valid ObjectId
     return res.status(404).send('Shift not found')
@@ -109,17 +135,31 @@ const deleteShift = async (req, res) => {
 
   // The following steps needs to be wraped in a try-catch block because .findByIdAndRemove will throw an error if a shiftId is not a valid ObjectId
   try {
-    // 2. Search for that shift in the database and delete it
-    const deletedShift = await Shift.findByIdAndRemove(shiftId)
+    let isDemo
+    if (req.user) { isDemo = req.user.isDemo }
+    if (isDemo) { // If the user is a demo, return a success response without updating the database
+      // 2. Search for that shift in the database
+      const deletedShift = await Shift.findById(shiftId)
 
-    // 3. If no shift is found, send back 404 error (resource not found)
-    if (!deletedShift) {
-      return res.status(404).send('Shift not found')
+      // 3. If no shift is found, send back 404 error (resource not found)
+      if (!deletedShift) {
+        return res.status(404).send('Shift not found')
+      }
+
+      // 4. Send back the deleted shift
+      return res.send(deletedShift)
+    } else {
+      // 2. Search for that shift in the database and delete it
+      const deletedShift = await Shift.findByIdAndRemove(shiftId)
+
+      // 3. If no shift is found, send back 404 error (resource not found)
+      if (!deletedShift) {
+        return res.status(404).send('Shift not found')
+      }
+
+      // 4. Send back the deleted shift
+      return res.send(deletedShift)
     }
-
-    // 4. Send back the deleted shift
-    res.send(deletedShift)
-
   } catch (error) {
     // An error will be thrown if shiftId is not a valid ObjectId
     return res.status(404).send('Shift not found')
@@ -130,22 +170,37 @@ const deleteShift = async (req, res) => {
 const pendingShifts = async (req, res) => {
   // I'm put everything in a try-catch block because I'm paranoid
   try {
+    let businessId = ''
     // 1. Extract business id from the jwt payload
-    const businessId = '5b53377c46556409ebbad3bc' // TODO: Change this to businessId = req.user.businessId after the authorize middleware has been added
+    if (process.env.NODE_ENV === 'development') {
+      let businesses = await Business.find()
+      businessId = businesses[0]._id // This is only to allow for development
+    } else {
+      businessId = req.user.businessId
+      console.log('works:', businessId)
+    }
 
     // 2. Search for all shifts that have that businessId
-    const allShifts = await Shift.find()
-      // .and([ { business: businessId }, { status: 'pending' } ])
-      .and([ { status: 'pending' } ]) // TODO: Hacky workaround: Remove this line, replace with one above
-      .populate('employee')
+    let allShifts = []
+    if (process.env.NODE_ENV === 'development') {
+      allShifts = await Shift.find()
+        .and([ { status: 'pending' } ]) // This is only to allow for development
+        .populate('employee')
+    } else {
+      allShifts = await Shift.find()
+        .and([ { business: businessId }, { status: 'pending' } ])
+        .populate('employee')
+    }
 
     // 3. If no shifts are found, send back 404 error (resource not found)
     if (allShifts.length === 0) {
       return res.status(404).allShiftssend('No Shifts Found')
     }
+
     // 4. Send back all the shifts
-    res.send(allShifts)
+    res.status(200).send(allShifts)
   } catch (error) {
+    console.log(error)
     res.status(500).send('Internal Server Error: (pendingShifts)')
   }
 }
@@ -156,7 +211,7 @@ const approveShift = async (req, res) => {
   try {
     // 1. Get the shift id from the URL params
     const shiftId = req.params.id
-  
+
     // 2. Find the shift in the database
     const shift = await Shift.findById(shiftId)
 
@@ -167,19 +222,27 @@ const approveShift = async (req, res) => {
 
     // 4. If shift status is not pending, return 403 error (not allowed to approve shift)
     if (shift.status !== 'pending') {
-      return res.status(403).send('Can\'t Update Shift Status') 
+      return res.status(403).send('Can\'t Update Shift Status')
     }
 
-    // 5. Update the shift status to approved
-    shift.set({ // 2. Update the movie
-      status: 'approved'
-    })
-    const updatedShift = await shift.save()
-
-    // 6. Send back the updated shift
-    res.send(updatedShift)
+    let isDemo
+    if (req.user) { isDemo = req.user.isDemo }
+    if (isDemo) { // If the user is a demo, return a success response without updating the database
+      shift.status = 'approved'
+      return res.status(200).send(shift)
+    } else {
+      // 5. Update the shift status to approved
+      shift.set({ // 2. Update the movie
+        status: 'approved'
+      })
+      const updatedShift = await shift.save()
+      
+      // 6. Send back the updated shift
+      return res.status(200).send(updatedShift)
+    }
   } catch (error) {
-    res.status(500).send('Something went wrong')
+    console.log(error)
+    res.status(500).send('Internal Server Error: ' + error)
   }
 }
 
@@ -187,29 +250,49 @@ const approveShift = async (req, res) => {
 const approveAllShifts = async (req, res) => {
   // I'm put everything in a try-catch block because I'm paranoid
   try {
+    let businessId = ''
     // 1. Get the business Id from the jwt payload
-    const businessId = '123' // TODO: Change this to businessId = req.user.businessId after the authorize middleware has been added
+    if (process.env.NODE_ENV === 'development') {
+      let businesses = await Business.find()
+      businessId = businesses[0]._id  // This is only to allow for development
+    } else {
+      businessId = req.user.businessId
+    }
 
     // 2. Search for all shifts that have that businessId
-    const allShifts = await Shift.find()
-      // .and([ { business: businessId }, { status: 'pending' } ])
-      .and([ { status: 'pending' } ]) // TODO: Hacky workaround: Remove this line, replace with one above
+    let allShifts = []
+    if (process.env.NODE_ENV === 'development') {
+      allShifts = await Shift.find()
+        .and([ { status: 'pending' } ]) // Delete? This is only to allow for development
+    } else {
+      allShifts = await Shift.find()
+        .and([ { business: businessId }, { status: 'pending' } ])
+    }
 
     // 3. If no shifts are found, send back 404 error (resource not found)
     if (allShifts.length === 0) {
       return res.status(404).send('No Shifts Found')
     }
 
-    // 4. Update the shift statuses to approved
-    await allShifts.forEach((shift) => {
-      shift.set({
-        status: 'approved'
+    let isDemo
+    if (req.user) { isDemo = req.user.isDemo }
+    if (isDemo) { // If the user is a demo, return a success response without updating the database
+      allShifts.forEach((shift) => {
+        shift.status = 'approved'
       })
-      shift.save()
-    })
+      return res.status(200).send('All Shifts Approved')
+    } else {
+      // 4. Update the shift statuses to approved
+      await allShifts.forEach((shift) => {
+        shift.set({
+          status: 'approved'
+        })
+        shift.save()
+      })
+    }
 
     // 5. Send back a confirmation message
-    return res.send('All Shifts Approved')
+    return res.status(200).send('All Shifts Approved')
   } catch (error) {
     res.status(500).send('Something went wrong')
   }
@@ -235,16 +318,25 @@ const rejectShift = async (req, res) => {
       return res.status(403).send('Can\'t Update Shift Status')
     }
 
-    // 5. Update the shift status to approved
-    shift.set({ // 2. Update the movie
-      status: 'rejected'
-    })
-    const updatedShift = await shift.save()
+    let isDemo
+    if (req.user) {
+      isDemo = req.user.isDemo
+    }
+    if (isDemo) { // If the user is a demo, return a success response without updating the database
+      shift.status = 'rejected'
+      return res.status(200).send(shift)
+    } else {
+      // 5. Update the shift status to approved
+      shift.set({ // 2. Update the movie
+        status: 'rejected'
+      })
+      const updatedShift = await shift.save()
 
-    // 6. Send back the updated shift
-    res.send(updatedShift)
+      // 6. Send back the updated shift
+      return res.status(200).send(updatedShift)
+    }
   } catch (error) {
-    res.status(500).send('Something went wrong')
+    res.status(500).send('Internal Server Error: ' + error)
   }
 }
 
@@ -252,13 +344,25 @@ const rejectShift = async (req, res) => {
 const getAllShifts = async (req, res) => {
   // I'm put everything in a try-catch block because I'm paranoid
   try {
+    let businessId = ''
     // 1. Extract business id from the jwt payload
-    const businessId = '5b53377c46556409ebbad3bc' // TODO: Change this to businessId = req.user.businessId after the authorize middleware has been added
+    if (process.env.NODE_ENV === 'development') {
+      let businesses = await Business.find()
+      businessId = businesses[0]._id  // This is only to allow for development
+    } else {
+      businessId = req.user.businessId
+    }
 
     // 2. Search for all shifts that have that businessId
-    const allShifts = await Shift.find()
-      // .and([ { business: businessId } ]) // TODO: Hacky workaround: Uncomment this line
-      .populate('employee')
+    let allShifts = []
+    if (process.env.NODE_ENV === 'development') {
+      allShifts = await Shift.find() // This is only to allow for development
+        .populate('employee')
+    } else {
+      allShifts = await Shift.find()
+        .and([ { business: businessId } ])
+        .populate('employee')
+    }
 
     // 3. If no shifts are found, send back 404 error (resource not found)
     if (allShifts.length === 0) {
@@ -273,6 +377,10 @@ const getAllShifts = async (req, res) => {
   }
 }
 
+const testRoute = (req, res) => {
+  res.status(200).json({ data: 'test' })
+}
+
 // export all controller functions required by router
 module.exports = {
   createShift,
@@ -283,5 +391,6 @@ module.exports = {
   approveShift,
   approveAllShifts,
   rejectShift,
-  getAllShifts
+  getAllShifts,
+  testRoute
 }
